most of functions is complete next correct fetching data from openweather in selected screen and calendar screen

+++ lib/library/weather/widget/hourly_winget.dart
@@ -96,7 +96,7 @@ class _HourlyDataWidgetState extends State<HourlyDataWidget> {
 }
 
 class HourlyDetails extends StatelessWidget {
-  final int tmp;
+  final double tmp;
   final int timeStamp;
   final String icon;
   final int index;
